Move invocation tests to own test file
